+++ schema/parser/parser.go
@@ -2,21 +2,22 @@ package parser
 
 import (
 	"text/scanner"
+	"unicode/utf8"
 
 	"github.com/alecthomas/participle/v2"
 	"github.com/alecthomas/participle/v2/lexer"
-	"github.com/teamkeel/keel/inputs"
+	"github.com/teamkeel/keel/model"
 	"github.com/teamkeel/keel/schema/expressions"
 )
 
 type Schema struct {
-	Pos lexer.Position
+	Node
 
 	Declarations []*Declaration `@@+`
 }
 
 type Declaration struct {
-	Pos lexer.Position
+	Node
 
 	Model *Model `("model" @@`
 	Role  *Role  `| "role" @@`
@@ -24,14 +25,14 @@ type Declaration struct {
 }
 
 type Model struct {
-	Pos lexer.Position
+	Node
 
-	Name     string          `@Ident`
+	Name     NameToken       `@@`
 	Sections []*ModelSection `"{" @@* "}"`
 }
 
 type ModelSection struct {
-	Pos lexer.Position
+	Node
 
 	Fields     []*ModelField  `( "fields" "{" @@+ "}"`
 	Functions  []*ModelAction `| "functions" "{" @@+ "}"`
@@ -39,93 +40,145 @@ type ModelSection struct {
 	Attribute  *Attribute     `| @@)`
 }
 
+type NameToken struct {
+	Node
+
+	Text string `@Ident`
+}
+
+type AttributeNameToken struct {
+	Node
+
+	Text string `"@" @Ident`
+}
+
 type ModelField struct {
-	Pos lexer.Position
+	Node
 
 	BuiltIn    bool
-	Name       string       `@Ident`
+	Name       NameToken    `@@`
 	Type       string       `@Ident`
 	Repeated   bool         `@( "[" "]" )?`
 	Attributes []*Attribute `( "{" @@+ "}" )?`
 }
 
 type API struct {
-	Pos lexer.Position
+	Node
 
-	Name     string        `@Ident`
+	Name     NameToken     `@@`
 	Sections []*APISection `"{" @@* "}"`
 }
 
 type Role struct {
-	Pos lexer.Position
+	Node
 
-	Name     string         `@Ident`
+	Name     NameToken      `@@`
 	Sections []*RoleSection `"{" @@* "}"`
 }
 
 type RoleSection struct {
-	Pos lexer.Position
+	Node
 
 	Domains []*RoleDomain `("domains" "{" @@* "}"`
 	Emails  []*RoleEmail  `| "emails" "{" @@* "}")`
 }
 
 type RoleDomain struct {
-	Pos lexer.Position
+	Node
 
 	Domain string `@String`
 }
 
 type RoleEmail struct {
-	Pos lexer.Position
+	Node
 
 	Email string `@String`
 }
 
 type APISection struct {
-	Pos lexer.Position
+	Node
 
 	Models    []*APIModels `("models" "{" @@* "}"`
 	Attribute *Attribute   `| @@)`
 }
 
 type APIModels struct {
-	Pos lexer.Position
+	Node
 
-	ModelName string `@Ident`
+	Name NameToken `@@`
 }
 
 type Attribute struct {
-	Pos lexer.Position
+	Node
 
-	Name      string               `"@" @Ident`
+	Name      AttributeNameToken   `@@`
 	Arguments []*AttributeArgument `( "(" @@ ( "," @@ )* ")" )?`
 }
 
 type AttributeArgument struct {
-	Pos lexer.Position
+	Node
 
-	Name       string                  `(@Ident ":")?`
+	Name       NameToken               `(@@ ":")?`
 	Expression *expressions.Expression `@@`
 }
 
 type ModelAction struct {
-	Pos lexer.Position
+	Node
 
 	Type       string       `@Ident`
-	Name       string       `@Ident`
+	Name       NameToken    `@@`
 	Arguments  []*ActionArg `"(" ( @@ ( "," @@ )* )? ")"`
 	Attributes []*Attribute `( "{" @@+ "}" )?`
 }
 
 type ActionArg struct {
-	Pos lexer.Position
+	Node
 
-	Name string `@Ident`
+	Name NameToken `@@`
 }
 
-func Parse(s *inputs.SchemaFile) (*Schema, error) {
+type Node struct {
+	Pos    lexer.Position
+	EndPos lexer.Position
+	Tokens []lexer.Token
+}
 
+// GetPositionRange returns a start and end position that correspond to Node
+// The behaviour of start position is exactly the same as the Pos field that
+// participle provides but the end position is calculated from the position of
+// the last token in this node, which is more useful if you want to know where
+// _this_ node starts and ends.
+func (n Node) GetPositionRange() (start lexer.Position, end lexer.Position) {
+	start.Column = n.Pos.Column
+	start.Filename = n.Pos.Filename
+	start.Line = n.Pos.Line
+	start.Offset = n.Pos.Offset
+
+	// This shouldn't really happen but just to be safe
+	if len(n.Tokens) == 0 {
+		return start, n.EndPos
+	}
+
+	lastToken := n.Tokens[len(n.Tokens)-1]
+	endPos := lastToken.Pos
+
+	tokenLength := utf8.RuneCountInString(lastToken.Value)
+
+	end.Filename = endPos.Filename
+
+	// assumption here is that a token can't span multiple lines, which
+	// I'm pretty sure is true
+	end.Line = endPos.Line
+
+	// Update offset and column to reflect the end of last token
+	// in this node
+	end.Offset = endPos.Offset + tokenLength
+	end.Column = endPos.Column + tokenLength
+
+	return start, end
+}
+
+func Parse(s *model.SchemaFile) (*Schema, error) {
 	// Customise the lexer to not ignore comments
 	lex := lexer.NewTextScannerLexer(func(s *scanner.Scanner) {
 		s.Mode =
@@ -143,7 +196,7 @@ func Parse(s *inputs.SchemaFile) (*Schema, error) {
 	}
 
 	schema := &Schema{}
-	// TODO: pass filename as first argument
+
 	err = parser.ParseString(s.FileName, s.Contents, schema)
 	if err != nil {
 		return nil, err